pyproject.toml
[tool.poetry]
name = "game-shop-django"
version = "0.1.0"
description = ""
authors = ["Pavlik-Ravlik <pastet1990@icloud.com>"]
readme = "README.md"
packages = [{include = "game_shop_django"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
psycopg2 = "^2.9.7"
pillow = "^10.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
